Split calc, plotting, checking to separate files.
