+++ python/lsst/validate/drp/checkAstrometryPhotometry.py
@@ -25,7 +25,6 @@ from __future__ import print_function, division
 import os.path
 import sys
 
-import matplotlib.pylab as plt
 import numpy as np
 import scipy.stats
 
@@ -38,12 +37,8 @@ import lsst.afw.geom as afwGeom
 import lsst.afw.image as afwImage
 import lsst.afw.coord as afwCoord
 
-# Plotting defaults
-plt.rcParams['axes.linewidth'] = 2
-plt.rcParams['mathtext.default'] = 'regular'
-plt.rcParams['font.size'] = 20
-plt.rcParams['axes.labelsize'] = 20
-# plt.rcParams['figure.titlesize'] = 30
+from .plotAstrometryPhotometry import plotAstrometry, plotPhotometryRms, plotPA1
+from .calcSrd import computeWidths, getRandomDiff, calcPA1, calcPA2
 
 def getCcdKeyName(dataid):
     """Return the key in a dataId that's referring to the CCD or moral equivalent.
@@ -207,22 +202,6 @@ def loadAndMatchData(repo, visitDataIds,
     return allMatches
 
 
-# The SRD recommends computing repeatability from a histogram of magnitude differences for the same star measured on two visits (using a median over the diffs to reject outliers). Since our dataset includes N>2 measurements for each star, we select a random pair of visits for each star.
-# We also divide each difference by sqrt(2), because we want the RMS about the (unknown) mean magnitude, not the RMS difference, and convert from mags to mmag.
-def getRandomDiff(array):
-    # not the most efficient way to extract a pair, but it's the easiest to write
-    copy = array.copy()
-    np.random.shuffle(copy)
-    return 1000*(copy[0] - copy[1])/(2**0.5)
-
-# We estimate the width of the histogram in two ways: using a simple RMS, and using the interquartile range (scaled by the IQR/RMS ratio for a Gaussian). We do this for 50 different random realizations of the measurement pairs, to provide some estimate of the uncertainty on our RMS estimates due to the random shuffling (we could probably turn this into a more formal estimate somehow, but I'm not going to bother with that at the moment).
-# While the SRD specifies that we should just compute the RMS directly, we haven't limited our sample to nonvariable stars as carefully as the SRD specifies, so using a more robust estimator like IQR will allow us to reject some outliers. It is also less sensitive some realistic sources of scatter the metric should include, however, such as bad zero points.
-
-def computeWidths(diffs):
-    rmsSigma = np.mean(diffs**2)**0.5
-    iqrSigma = np.subtract.reduce(np.percentile(diffs, [75, 25])) / (scipy.stats.norm.ppf(0.75)*2)
-    return rmsSigma, iqrSigma
-
 
 def analyzeData(allMatches):
     """Calculate summary statistics for each star.
@@ -302,59 +281,6 @@ def analyzeData(allMatches):
     return info_struct, safeMatches
 
 
-def plot_medians(ax, x1, x2, x1_color='blue', x2_color='red'):
-    ax.axhline(x1, color='white', linewidth=4)
-    ax.axhline(x2, color='white', linewidth=4)
-    ax.axhline(x1, color=x1_color, linewidth=3)
-    ax.axhline(x2, color=x2_color, linewidth=3)
-
-
-def plotAstrometry(mag, mmagrms, dist, match, good_mag_limit=19.5):
-    """Plot angular distance between matched sources from different exposures.
-
-    @param[in] mag    Magnitude.  List or numpy.array.
-    @param[in] mmagrms    Magnitude RMS.  List or numpy.array.
-    @param[in] dist   Separation from reference.  List of numpy.array
-    @param[in] match  Number of stars matched.  Integer.
-    """
-
-    bright, = np.where(np.asarray(mag) < good_mag_limit)
-
-    dist_median = np.median(dist) 
-    bright_dist_median = np.median(np.asarray(dist)[bright])
-
-    fig, ax = plt.subplots(ncols=2, nrows=1, figsize=(18, 12))
-
-    ax[0].hist(dist, bins=100, color='blue',
-               histtype='stepfilled', orientation='horizontal')
-    ax[0].hist(np.asarray(dist)[bright], bins=100, color='red',
-               histtype='stepfilled', orientation='horizontal',
-               label='mag < %.1f' % good_mag_limit)
-
-    ax[0].set_ylim([0., 500.])
-    ax[0].set_ylabel("Distance in mas")
-    ax[0].set_title("Median : %.1f, %.1f mas" % 
-                       (bright_dist_median, dist_median),
-                       x=0.55, y=0.88)
-    plot_medians(ax[0], dist_median, bright_dist_median)
-
-    ax[1].scatter(mag, dist, s=10, color='blue', label='All')
-    ax[1].scatter(np.asarray(mag)[bright], np.asarray(dist)[bright], s=10, 
-                  color='red', 
-                  label='mag < %.1f' % good_mag_limit)
-    ax[1].set_xlabel("Magnitude")
-    ax[1].set_ylabel("Distance in mas")
-    ax[1].set_xlim([17, 24])
-    ax[1].set_ylim([0., 500.])
-    ax[1].set_title("# of matches : %d, %d" % (len(bright), match))
-    ax[1].legend(loc='upper left')
-    plot_medians(ax[1], dist_median, bright_dist_median)
-
-    plt.suptitle("Astrometry Check", fontsize=30)
-    plotPath = "check_astrometry.png"
-    plt.savefig(plotPath, format="png")
-
-
 def checkAstrometry(mag, mmagrms, dist, match,
                     good_mag_limit=19.5,
                     medianRef=100, matchRef=500):
@@ -426,89 +352,12 @@ def checkPhotometry(mag, mmagrms, dist, match,
 
     return photoScatter
 
-def plotPhotometryRms(mag, mmagrms, dist, match, good_mag_limit=19.5):
-    """Plot photometric RMS for matched sources.
 
-    @param[in] mag    Magnitude.  List or numpy.array.
-    @param[in] mmagrms    Magnitude RMS.  List or numpy.array.
-    @param[in] dist   Separation from reference.  List of numpy.array
-    @param[in] match  Number of stars matched.  Integer.
-    """
-
-    bright, = np.where(np.asarray(mag) < good_mag_limit)
-
-    mmagrms_median = np.median(mmagrms) 
-    bright_mmagrms_median = np.median(np.asarray(mmagrms)[bright])
-
-    fig, ax = plt.subplots(ncols=2, nrows=1, figsize=(18, 12))
-    ax[0].hist(mmagrms, bins=100, range=(0, 500), color='blue', label='All',
-                  histtype='stepfilled', orientation='horizontal')
-    ax[0].hist(np.asarray(mmagrms)[bright], bins=100, range=(0, 500), color='red', 
-                  label='mag < %.1f' % good_mag_limit,
-                  histtype='stepfilled', orientation='horizontal')
-    ax[0].set_ylim([0, 500])
-    ax[0].set_ylabel("RMS in mmag")
-    ax[0].set_title("Median : %.1f, %.1f mmag" % 
-                    (bright_mmagrms_median, mmagrms_median),
-                    x=0.55, y=0.88)
-    plot_medians(ax[0], mmagrms_median, bright_mmagrms_median)
-
-    ax[1].scatter(mag, mmagrms, s=10, color='blue', label='All')
-    ax[1].scatter(np.asarray(mag)[bright], np.asarray(mmagrms)[bright], 
-                     s=10, color='red', 
-                     label='mag < %.1f' % good_mag_limit)
-
-    ax[1].set_xlabel("Magnitude")
-    ax[1].set_ylabel("RMS [mmag]")
-    ax[1].set_xlim([17, 24])
-    ax[1].set_ylim([0, 500])
-    ax[1].set_title("# of matches : %d, %d" % (len(bright), match))
-    ax[1].legend(loc='upper left')
-    plot_medians(ax[1], mmagrms_median, bright_mmagrms_median)
-
-    plt.suptitle("Photometry Check", fontsize=30)
-    plotPath = "check_photometry.png"
-    plt.savefig(plotPath, format="png")
-
-
-# To validate these width estimates, we can plot a single random realization (re-evaluate the cell or call the function again to get a new one).
-def plotPA1(gv, psfMagKey):
-#    diffs = gv.aggregate(getRandomDiff, field=apMagKey)
-#    means = gv.aggregate(np.mean, field=apMagKey)
-    diffs = gv.aggregate(getRandomDiff, field=psfMagKey)
-    means = gv.aggregate(np.mean, field=psfMagKey)
-    rmsPA1, iqrPA1 = computeWidths(diffs)
-    fig = plt.figure(figsize=(18,12))
-    ax1 = fig.add_subplot(1,2,1)
-    ax1.scatter(means, diffs, s=8, linewidth=0, alpha=0.4)
-    ax1.axhline(rmsPA1, color='r')
-    ax1.axhline(-rmsPA1, color='r')
-    ax1.axhline(iqrPA1, color='g')
-    ax1.axhline(-iqrPA1, color='g')
-    ax2 = fig.add_subplot(1,2,2, sharey=ax1)
-    ax2.hist(diffs, bins=50, orientation='horizontal', normed=True, alpha=0.5)
-    yv = np.linspace(-100, 100, 100)
-    ax2.plot(scipy.stats.norm.pdf(yv, scale=rmsPA1), yv, 'r-', label="PA1(RMS)=%4.2f mmag" % rmsPA1)
-    ax2.plot(scipy.stats.norm.pdf(yv, scale=iqrPA1), yv, 'g-', label="PA1(IQR)=%4.2f mmag" % iqrPA1)
-    ax2.set_ylim(-100, 100)
-    ax2.legend()
-#    ax1.set_ylabel(u"12-pixel aperture magnitude diff (mmag)")
-#    ax1.set_xlabel(u"12-pixel aperture magnitude")
-    ax1.set_xlabel("psf aperture magnitude")
-    ax1.set_ylabel("psf magnitude diff (mmag)")
-    for label in ax2.get_yticklabels(): label.set_visible(False)
-
-    plt.savefig("PA1.png")
-
-
-# The PF1/PA2 puts a limit on the fraction of outliers in the above plot. Below, we compute the values of PA2 for the minimum, design, and stretch specification values of PF1 in the SRD.
-def printPA2(gv, psfMagKey):
-#    diffs = gv.aggregate(getRandomDiff, field=apMagKey)
-    diffs = gv.aggregate(getRandomDiff, field=psfMagKey)
-    minPA2, designPA2, stretchPA2 = np.percentile(np.abs(diffs), [80, 90, 95])
+def printPA2(gv, magKey):
+    minPA2, designPA2, stretchPA2 = calcPA2(gv, magKey)
     print("minimum: PF1=20%% of diffs more than PA2=%4.2f mmag (target is PA2 < 15 mmag)" % minPA2)
-    print("design: PF1=10%% of diffs more than PA2=%4.2f mmag (target is PA2 < 15 mmag)" % designPA2)
-    print("stretch: PF1=5%% of diffs more than PA2=%4.2f mmag (target is PA2 < 10 mmag)" % stretchPA2)
+    print("design:  PF1=10%% of diffs more than PA2=%4.2f mmag (target is PA2 < 15 mmag)" % designPA2)
+    print("stretch: PF1= 5%% of diffs more than PA2=%4.2f mmag (target is PA2 < 10 mmag)" % stretchPA2)
 
 ####
 def run(repo, visitDataIds, good_mag_limit, 
@@ -531,7 +380,6 @@ def run(repo, visitDataIds, good_mag_limit,
                     medianRef=medianPhotoscatterRef, matchRef=matchRef)
     plotAstrometry(magavg, mmagrms, dist, match, good_mag_limit=good_mag_limit)
     plotPhotometryRms(magavg, mmagrms, dist, match, good_mag_limit=good_mag_limit)
-#    plotPhotometryRms(magavg, delta_mag, dist, match, good_mag_limit=good_mag_limit)
 
     psfMagKey = allMatches.schema.find("base_PsfFlux_mag").key
     plotPA1(safeMatches, psfMagKey)
